I've refactored the Day9 route calculation and parsing. Here's what I did:
- I moved the Day9 and Route classes to a new 'day9' package (tar.calion.aoc.year2015.day9).
- I extracted the input parsing logic from the Day9 constructor into a new private method called 'parseRoutesInput'.
- I introduced a new 'Routes' record to wrap the calculated shortest route.
- I renamed the 'calculateShortestRoute' method in Day9 to 'calculateRoutes' and changed its return type to 'Routes'.
- I updated Day9Test.java to reflect these changes. This included moving the test file to the new package structure and adjusting the test logic for the new method signature and return type.
