+++ src/main/java/tar/calion/aoc/year2015/day9/Routes.java
@@ -1,0 +1,6 @@
+package tar.calion.aoc.year2015.day9;
+
+import tar.calion.aoc.year2015.day9.Route;
+
+public record Routes(Route shortestRoute) {
+}
